backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-dotenv
python-multipart
